requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14,<4.0
psycopg2>=2.9,<3.0
Pillow>=9.0,<10.0
markdown>=3.3,<4.0
django-filter>=22.1,<23.0
djangorestframework-simplejwt>=5.2,<6.0
django-cors-headers>=3.13,<4.0